+++ test/harness/arm_float16/test.c
@@ -1,0 +1,6 @@
+/* extended floating point types */
+static void f(void)
+{
+    __fp16 a1 = 2.0;
+    __bf16 b1 = 2.0;
+}